requirements.txt
Flask==3.0.0
boto3==1.34.0
hedera-sdk-python==2.34.0
python-dotenv==1.0.0
requests
itsdangerous==2.2.0
Jinja2==3.1.6
Werkzeug==3.1.3
MarkupSafe==3.0.2
gunicorn==23.0.0